+++ packages/generator/src/index.ts
@@ -2,6 +2,7 @@ export * from "./generators/app-generator"
 export * from "./generators/model-generator"
 export * from "./generators/mutation-generator"
 export * from "./generators/page-generator"
+export * from "./generators/queries-generator"
 export * from "./generators/query-generator"
 export * from "./generators/form-generator"
 export * from "./generator"
