+++ packages/generator/templates/queries/get__ModelName__.ts
