+++ packages/generator/templates/queries/get__ModelNames__.ts
