+++ packages/generator/templates/query/__modelName__.ts
@@ -1,0 +1,8 @@
+import db from "db"
+
+export default async function __modelName__(
+  __,
+  ctx: Record<any, any> = {},
+) {
+  return ctx
+}
